+++ src/components/collection/collection-item.styl
